chat-app-backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
passlib[bcrypt]
python-jose
requests
python-dotenv
email-validator
bcrypt==4.0.1
sentence-transformers
numpy